--- a/benchmark/benchmarks.jl
+++ b/benchmark/benchmarks.jl
@@ -16,7 +16,7 @@ tst_types = (Gray{N0f8}, Gray{Float32})
 
 const SUITE = BenchmarkGroup()
 
-alg_list = (("Original", lbp_original),
+alg_list = (("Original", local_binary_pattern),
             ("Multi-block (1x1)", img->multiblock_lbp(img, (1, 1))),
             ("Multi-block (3x3)", img->multiblock_lbp(img, (3, 3))),
             ("Multi-block (5x5)", img->multiblock_lbp(img, (5, 5))),