--- a/docs/src/reference.md
+++ b/docs/src/reference.md
@@ -1,7 +1,7 @@
 # API References
 
 ```@docs
-lbp_original
+local_binary_pattern
 ```
 
 ## Helpers