--- a/src/local_binary_pattern.jl
+++ b/src/local_binary_pattern.jl
@@ -1,7 +1,32 @@
-"""
-    lbp_original([f], X; [rotation], [uniform_degree])
+###
+# Various LBP modes
+###
 
-Compute the local binary pattern of gray image `X` using the original method.
+abstract type LocalBinaryPattern end
+
+struct LBP{F} <: LocalBinaryPattern
+    mode::F # mode(X::AbstractArray{T,N}, I::CartesianIndex{N}, offsets::Tuple)
+end
+
+Base.@propagate_inbounds _default_lbp_mode(X, I, offsets) = X[I]
+
+Base.@propagate_inbounds function lbp_encode(lbp::LBP, X::AbstractMatrix, I::CartesianIndex{2}, offsets::Tuple)
+    rst = 0
+    gc = lbp.mode(X, I, offsets)
+    for i in 1:length(offsets)
+        p = I.I .+ offsets[i]
+        @boundscheck checkbounds(Bool, X, p...) || continue
+        rst += ifelse(gc <= _inbounds_getindex(X, p), 1, 0) << (i-1)
+    end
+    return rst
+end
+
+
+# non-interpolation version
+"""
+    local_binary_pattern([f], X; [rotation], [uniform_degree])
+
+Compute the local binary pattern of gray image `X`.
 
 # Arguments
 
@@ -33,19 +58,19 @@ julia> X = [6 7 9; 5 6 3; 2 1 7]
  5  6  3
  2  1  7
 
-julia> lbp_original(X)
+julia> local_binary_pattern(X)
 3×3 $(Matrix{UInt8}):
  0xc0  0x40  0x00
  0x68  0xa9  0x1b
  0x28  0x6b  0x00
 
-julia> lbp_original(X; rotation=true)
+julia> local_binary_pattern(X; rotation=true)
 3×3 $(Matrix{UInt8}):
  0x03  0x01  0x00
  0x0d  0x35  0x1b
  0x05  0x5b  0x00
 
-julia> lbp_original(X; uniform_degree=2)
+julia> local_binary_pattern(X; uniform_degree=2)
 3×3 $(Matrix{UInt8}):
  0xc0  0x40  0x00
  0x09  0x09  0x09
@@ -100,23 +125,25 @@ unchanged because it only has ``2`` bit transitions.
 - [3] Pietikäinen, Matti, Timo Ojala, and Zelin Xu. "Rotation-invariant texture classification using feature distributions." _Pattern recognition_ 33.1 (2000): 43-52.
 - [4] T. Ojala, M. Pietikainen, and T. Maenpaa, “Multiresolution gray-scale and rotation invariant texture classification with local binary patterns,” _IEEE Trans. Pattern Anal. Machine Intell._, vol. 24, no. 7, pp. 971–987, Jul. 2002, doi: 10.1109/TPAMI.2002.1017623.
 """
-function lbp_original(f, X::AbstractArray; rotation::Bool=false, uniform_degree::Union{Nothing,Int}=nothing)
+function local_binary_pattern(f, X::AbstractArray; rotation::Bool=false, uniform_degree::Union{Nothing,Int}=nothing)
     # The original version [1] uses clockwise order; here we use anti-clockwise order
     # because Julia is column-major order. If we consider memory order differences then they
     # are equivalent.
     offsets = ((-1, -1), (0, -1), (1, -1), (-1, 0), (1, 0), (-1, 1), (0, 1), (1, 1))
-    lookups = _build_lbp_original_lookup(UInt8, 8; rotation=rotation, uniform_degree=uniform_degree)
-    lbp_original!(similar(X, UInt8), f, X, offsets, lookups)
+    lookups = _build_local_binary_pattern_lookup(UInt8, 8; rotation=rotation, uniform_degree=uniform_degree)
+    local_binary_pattern!(similar(X, UInt8), LBP(f), X, offsets, lookups)
 end
+local_binary_pattern(X::AbstractArray; kwargs...) = local_binary_pattern(_default_lbp_mode, X; kwargs...)
 
+# interpolation version
 """
-    lbp_original([f], X, npoints, radius, interpolation=Linear(); [rotation], [uniform_degree])
+    local_binary_pattern([f], X, npoints, radius, interpolation=Linear(); [rotation], [uniform_degree])
 
 Compute the local binary pattern of gray image `X` using the interpolation-based original
 method with circular neighborhood matrix.
 
 This produces better result for `rotation=true` case but is usually slower than the plain
-3x3 matrix version `lbp_original(X)`.
+3x3 matrix version `local_binary_pattern(X)`.
 
 # Arguments
 
@@ -142,8 +169,8 @@ This produces better result for `rotation=true` case but is usually slower than 
 !!! note "neighborhood order differences"
     Different implementation might use different neighborhood orders; this will change the
     encoding result but will not change the overall distribution. For instance,
-    `lbp_original(X)` differs from `lbp_original(X, 8, 1, Constant())` only by how
-    `offsets` (see below) are ordered; the former uses column-major top-left to
+    `local_binary_pattern(X)` differs from `local_binary_pattern(X, 8, 1, Constant())` only
+    by how `offsets` (see below) are ordered; the former uses column-major top-left to
     bottom-right 3x3 matrix order and the latter uses circular order.
 
 # Examples
@@ -155,13 +182,13 @@ julia> X = [6 7 9; 5 6 3; 2 1 7]
  5  6  3
  2  1  7
 
-julia> lbp_original(X, 4, 1) # 4-neighbor with circular radius 1
+julia> local_binary_pattern(X, 4, 1) # 4-neighbor with circular radius 1
 3×3 $(Matrix{UInt32}):
  0x00000001  0x00000001  0x00000000
  0x00000003  0x00000002  0x0000000e
  0x00000002  0x00000007  0x00000000
 
-julia> lbp_original(X, 4, 1; rotation=true)
+julia> local_binary_pattern(X, 4, 1; rotation=true)
 3×3 $(Matrix{UInt32}):
  0x00000001  0x00000001  0x00000000
  0x00000003  0x00000001  0x00000007
@@ -172,12 +199,13 @@ julia> lbp_original(X, 4, 1; rotation=true)
 
 - [1] T. Ojala, M. Pietikainen, and T. Maenpaa, “Multiresolution gray-scale and rotation invariant texture classification with local binary patterns,” _IEEE Trans. Pattern Anal. Machine Intell._, vol. 24, no. 7, pp. 971–987, Jul. 2002, doi: 10.1109/TPAMI.2002.1017623.
 """
-function lbp_original(
+function local_binary_pattern(
         f, X::AbstractArray, npoints::Int, radius::Real, interpolation=Linear();
         rotation::Bool=false, uniform_degree::Union{Nothing,Int}=nothing)
     interpolation = wrap_BSpline(interpolation)
     offsets = _circular_neighbor_offsets(npoints, radius)
     if interpolation == BSpline(Constant())
+        # performance optimization: skip interpolation when it is nearest neighbor interpolation
         offsets = map(offsets) do o
             round.(Int, o, RoundToZero)
         end
@@ -186,22 +214,15 @@ function lbp_original(
         itp = interpolate(X, interpolation)
     end
     # For the sake of type-stability, hardcode to UInt32 here.
-    lookups = _build_lbp_original_lookup(UInt32, npoints, rotation=rotation, uniform_degree=uniform_degree)
-    lbp_original!(similar(X, UInt32), f, itp, offsets, lookups)
+    lookups = _build_local_binary_pattern_lookup(UInt32, npoints, rotation=rotation, uniform_degree=uniform_degree)
+    local_binary_pattern!(similar(X, UInt32), LBP(f), itp, offsets, lookups)
+end
+function local_binary_pattern(X::AbstractArray, npoints::Int, radius::Real, interpolation=Linear(); kwargs...)
+    local_binary_pattern(_default_lbp_mode, X, npoints, radius, interpolation; kwargs...)
 end
 
-lbp_original(X::AbstractArray; kwargs...) = lbp_original(_default_lbp_f, X; kwargs...)
-function lbp_original(X::AbstractArray, npoints::Int, radius::Real, interpolation=Linear(); kwargs...)
-    lbp_original(_default_lbp_f, X, npoints, radius, interpolation; kwargs...)
-end
 
-function lbp_original!(
-        out,
-        f,
-        X::AbstractMatrix{T},
-        offsets::Tuple,
-        lookups::Vector
-        ) where T<:Union{Real,Gray}
+function local_binary_pattern!(out::AbstractArray, lbp::LocalBinaryPattern, X::AbstractArray, offsets, lookups)
     length(offsets) > 32 && throw(ArgumentError("length(offsets)=$(length(offsets)) is expected to be no larger than 32."))
     outerR = CartesianIndices(X)
     r = CartesianIndex(
@@ -211,39 +232,25 @@ function lbp_original!(
     innerR = first(outerR)+r:last(outerR)-r
 
     # TODO(johnnychen94): use LoopVectorization
-    @inbounds for I in innerR
-        gc = f(X, I, offsets)
-        # This inner loop fuses the binary pattern build stage and encoding stage for
-        # better performance.
-        rst = 0
-        for i in 1:length(offsets)
-            p = I.I .+ offsets[i]
-            rst += ifelse(gc <= _inbounds_getindex(X, p), 1, 0) << (i-1)
-        end
-        out[I] = rst
+    @inbounds @simd for I in innerR
+        out[I] = lbp_encode(lbp, X, I, offsets)
     end
-
-    # boundary conditions are undefined in [1]; here we directly skip out-of-boundary values
-    @inbounds for I in EdgeIterator(outerR, innerR)
-        # this requires `f` to not propagate @inbounds unless it is plain `X[I]`
-        gc = f(X, I, offsets)
-        rst = 0
-        for i in 1:length(offsets)
-            p = I.I .+ offsets[i]
-            checkbounds(Bool, X, p...) || continue
-            rst += ifelse(gc <= _inbounds_getindex(X, p), 1, 0) << (i-1)
-        end
-        out[I] = rst
+    for I in EdgeIterator(outerR, innerR)
+        out[I] = lbp_encode(lbp, X, I, offsets)
     end
 
     for F in lookups
         @. out = F[out]
     end
-
     return out
 end
 
-function _build_lbp_original_lookup(
+
+###
+# Utils
+###
+
+function _build_local_binary_pattern_lookup(
         ::Type{T}, nbits;
         rotation::Bool,
         uniform_degree::Union{Nothing,Int}
@@ -260,11 +267,9 @@ function _build_lbp_original_lookup(
     return lookups
 end
 
-Base.@propagate_inbounds _default_lbp_f(X, I, offsets) = X[I]
-
 """
     average_mode(X, I, offsets) -> value
-    lbp_original(average_mode, X, args...; kwargs...)
+    local_binary_pattern(average_mode, X, args...; kwargs...)
 
 Compute the local binary pattern using the average mode of the block.
 
@@ -280,12 +285,14 @@ julia> X = [6 7 9; 5 6 3; 2 1 7]
  5  6  3
  2  1  7
 
-julia> lbp_original(average_mode, X)
+julia> local_binary_pattern(average_mode, X)
 3×3 $(Matrix{UInt8}):
  0xc0  0x40  0x00
  0x68  0xa9  0x1b
  0x28  0x69  0x01
 ```
+
+See also [`local_binary_pattern`](@ref) for additional arguments and parameters.
 """
 Base.@propagate_inbounds function average_mode(X, I::CartesianIndex, offsets)
     v = X[I]