--- a/src/local_binary_pattern.jl
+++ b/src/local_binary_pattern.jl
@@ -4,9 +4,15 @@
 
 abstract type LocalBinaryPattern end
 
-struct LBP{F} <: LocalBinaryPattern
+struct LBP{N,F} <: LocalBinaryPattern
     mode::F # mode(X::AbstractArray{T,N}, I::CartesianIndex{N}, offsets::Tuple)
+    block_size::Dims{N}
+    function LBP(mode::F, block_size::Dims{N}) where {N,F}
+        all(block_size .> 0) || throw(ArgumentError("block size `block_size=$(block_size)` should be positive integers."))
+        new{N, F}(mode, block_size)
+    end
 end
+LBP{N}(mode, block_size::Int=1) where N = LBP(mode, ntuple(_->block_size, N))
 
 Base.@propagate_inbounds _default_lbp_mode(X, I, offsets) = X[I]
 
@@ -17,6 +23,41 @@ Base.@propagate_inbounds function lbp_encode(lbp::LBP, X::AbstractMatrix, I::Car
         p = I.I .+ offsets[i]
         @boundscheck checkbounds(Bool, X, p...) || continue
         rst += ifelse(gc <= _inbounds_getindex(X, p), 1, 0) << (i-1)
+    end
+    return rst
+end
+
+
+Base.@propagate_inbounds function mlbp_encode_inbounds(
+        lbp::LBP,
+        iX::IntegralArray,
+        I::CartesianIndex{2},
+        bo::CartesianIndex{2},
+        offsets::Tuple)
+    rst = 0
+    gc = iX[I..I+bo]
+    for i in 1:length(offsets)
+        p = CartesianIndex(I.I .+ offsets[i])
+        rst += ifelse(gc <= iX[p..p+bo], 1, 0) << (i-1)
+    end
+    return rst
+end
+function mlbp_encode(
+        lbp::LBP,
+        iX::IntegralArray,
+        I::CartesianIndex{2},
+        bo::CartesianIndex{2},
+        offsets::Tuple)
+    rst = 0
+    Rlast = last(CartesianIndices(iX))
+    R = I:min(I+bo, Rlast)
+    gc = iX[first(R)..last(R)]/length(R)
+    for i in 1:length(offsets)
+        p = CartesianIndex(I.I .+ offsets[i])
+        checkbounds(Bool, iX, p.I...) || continue
+        Rp = p:min(p+bo, Rlast)
+        gp = iX[first(Rp)..last(Rp)]/length(Rp)
+        rst += ifelse(gc <= gp, 1, 0) << (i-1)
     end
     return rst
 end
@@ -131,7 +172,7 @@ function local_binary_pattern(f, X::AbstractArray; rotation::Bool=false, uniform
     # are equivalent.
     offsets = ((-1, -1), (0, -1), (1, -1), (-1, 0), (1, 0), (-1, 1), (0, 1), (1, 1))
     lookups = _build_local_binary_pattern_lookup(UInt8, 8; rotation=rotation, uniform_degree=uniform_degree)
-    local_binary_pattern!(similar(X, UInt8), LBP(f), X, offsets, lookups)
+    local_binary_pattern!(similar(X, UInt8), LBP{ndims(X)}(f), X, offsets, lookups)
 end
 local_binary_pattern(X::AbstractArray; kwargs...) = local_binary_pattern(_default_lbp_mode, X; kwargs...)
 
@@ -215,28 +256,125 @@ function local_binary_pattern(
     end
     # For the sake of type-stability, hardcode to UInt32 here.
     lookups = _build_local_binary_pattern_lookup(UInt32, npoints, rotation=rotation, uniform_degree=uniform_degree)
-    local_binary_pattern!(similar(X, UInt32), LBP(f), itp, offsets, lookups)
+    local_binary_pattern!(similar(X, UInt32), LBP{ndims(X)}(f), itp, offsets, lookups)
 end
 function local_binary_pattern(X::AbstractArray, npoints::Int, radius::Real, interpolation=Linear(); kwargs...)
     local_binary_pattern(_default_lbp_mode, X, npoints, radius, interpolation; kwargs...)
 end
 
+# Multi-blocks version
+"""
+    local_binary_pattern([f], X, block_size; [rotation], [uniform_degree])
 
-function local_binary_pattern!(out::AbstractArray, lbp::LocalBinaryPattern, X::AbstractArray, offsets, lookups)
+Compute the local binary pattern of gray image `X` of image blocks with size `block_size`.
+This method is usually called MB-LBP[1] or Locally Assembled Binary (LAB) feature[2] in the
+literature.
+
+# Arguments
+
+- `X::AbstractMatrix`: the input image matrix. For colorful images, one can manually convert
+  it to some monochrome space, e.g., `Gray`, the L-channel of `Lab`. One could also do
+  channelwise LBP and then concatenate together.
+- `block_size::Tuple{Int,Int}`: positive integers that used to specify the size of each
+  block. When `block_size == (1, 1)`, this method degenerates to the pixel-version
+  [`local_binary_pattern`](@ref).
+
+For the meaning and values of parameters `rotation` and `uniform_degree` please see the docs
+of [`local_binary_pattern`](@ref).
+
+# Examples
+
+```jldoctest; setup=:(using LocalBinaryPatterns)
+julia> X = zeros(Int, 9, 9); X[1:3, 1:3] .= 50; X[4:6, 4:6] .= 1; X[7:9, 7:9] .= 50; X
+9×9 $(Matrix{Int}):
+ 50  50  50  0  0  0   0   0   0
+ 50  50  50  0  0  0   0   0   0
+ 50  50  50  0  0  0   0   0   0
+  0   0   0  1  1  1   0   0   0
+  0   0   0  1  1  1   0   0   0
+  0   0   0  1  1  1   0   0   0
+  0   0   0  0  0  0  50  50  50
+  0   0   0  0  0  0  50  50  50
+  0   0   0  0  0  0  50  50  50
+
+julia> local_binary_pattern(X, (3, 3))[4, 4] # 0b1000_0001
+0x81
+
+julia> local_binary_pattern(X, (3, 3); rotation=true)[4, 4] # 0b0000_0011
+0x03
+
+julia> local_binary_pattern(X, (3, 3); uniform_degree=2)[4, 4] # 0x09 (i.e., miscellaneous pattern)
+0x09
+```
+
+# Extended help
+
+The following is how the block-version of local binary pattern computed with `block_size=(3,
+3)`. Check out [`local_binary_pattern`](@ref) for more details of the original pixel-version of
+local binary pattern.
+
+```text
+2  3  2  |  7  6  6  |  2  1  3
+1  2  3  |  5  3  4  |  7  7  3
+6  5  4  |  1  4  5  |  1  5  8
+-------------------------------     sum                compare          weighted          sum
+1  2  3  |  1  7  3  |  8  3  7            28  41  37           0  1  0           0 8 0
+1  2  1  |  4  3  6  |  5  2  2     ==>    20  38  45    ==>    0  x  1     ==>   0 x 64  ==> 200
+1  2  7  |  5  3  6  |  4  7  7            33  31  39           0  0  1           0 0 128
+-------------------------------
+4  2  4  |  8  1  5  |  3  7  7
+5  7  4  |  1  1  1  |  3  4  3
+3  3  1  |  5  1  8  |  1  8  3
+```
+
+# References
+
+- [1] Zhang, Lun, et al. "Face detection based on multi-block lbp representation." _International conference on biometrics_. Springer, Berlin, Heidelberg, 2007.
+- [2] Yan, Shengye, et al. "Locally assembled binary (LAB) feature with feature-centric cascade for fast and accurate face detection." _2008 IEEE Conference on Computer Vision and Pattern Recognition_. IEEE, 2008.
+"""
+function local_binary_pattern(f, X::AbstractMatrix, block_size::Dims{2}; rotation::Bool=false, uniform_degree::Union{Nothing,Int}=nothing)
+    offsets = ((-1, -1), (0, -1), (1, -1), (-1, 0), (1, 0), (-1, 1), (0, 1), (1, 1))
+    lookups = _build_local_binary_pattern_lookup(UInt8, 8; rotation=rotation, uniform_degree=uniform_degree)
+    local_binary_pattern!(similar(X, UInt8), LBP(f, block_size), X, offsets, lookups)
+end
+local_binary_pattern(X::AbstractMatrix, block_size::Dims{2}; kwargs...) = local_binary_pattern(_default_lbp_mode, X, block_size; kwargs...)
+
+function local_binary_pattern!(out::AbstractMatrix, lbp::LocalBinaryPattern, X::AbstractMatrix, offsets, lookups)
     length(offsets) > 32 && throw(ArgumentError("length(offsets)=$(length(offsets)) is expected to be no larger than 32."))
+
+    block_size = lbp.block_size
+    offsets = map(offsets) do o
+        o .* block_size
+    end
+    bo = CartesianIndex(block_size .- 1)
+
     outerR = CartesianIndices(X)
     r = CartesianIndex(
         ceil(Int, maximum(abs.(extrema(first.(offsets))))),
         ceil(Int, maximum(abs.(extrema(last.(offsets)))))
     )
-    innerR = first(outerR)+r:last(outerR)-r
+    innerR = first(outerR)+r:last(outerR)-r-bo
 
-    # TODO(johnnychen94): use LoopVectorization
-    @inbounds @simd for I in innerR
-        out[I] = lbp_encode(lbp, X, I, offsets)
-    end
-    for I in EdgeIterator(outerR, innerR)
-        out[I] = lbp_encode(lbp, X, I, offsets)
+    if block_size == (1, 1)
+        # TODO(johnnychen94): use LoopVectorization
+        @inbounds @simd for I in innerR
+            out[I] = lbp_encode(lbp, X, I, offsets)
+        end
+        for I in EdgeIterator(outerR, innerR)
+            out[I] = lbp_encode(lbp, X, I, offsets)
+        end
+    else
+        # It's computationally more efficient to use integral array to compute
+        # the mean block value.
+        iX = IntegralArray(X)
+        # TODO(johnnychen94): it can be faster if we can enable SIMD here.
+        #                     Need to skip the boundary check in IntegralArray.
+        @inbounds for I in innerR
+            out[I] = mlbp_encode_inbounds(lbp, iX, I, bo, offsets)
+        end
+        for I in EdgeIterator(outerR, innerR)
+            out[I] = mlbp_encode(lbp, iX, I, bo, offsets)
+        end
     end
 
     for F in lookups