This PR consists of several commits, the purpose is to reuse the `lbp_original` implementation for multiblock LBP and future LBP variants.

Since we don't limit it to the original version, the name `lbp_original` is not appropriate anymore; for this reason, I've renamed it to `local_binary_pattern`. If users think it's too long to type they can always write `const lbp = local_binary_pattern` in their scripts.

A perhaps surprising outcome is that the algorithms run more efficient:

```julia
img = rand(Float32, 512, 512);

@btime local_binary_pattern($img);
# PR: 656.618 μs (3 allocations: 256.16 KiB)
# master: 1.321 ms (3 allocations: 256.16 KiB)
```

<details>
  <summary> benchmark report (PR) </summary>

# Benchmark Report for *LocalBinaryPatterns*

## Job Properties
* Time of benchmark: 31 Dec 2021 - 4:56
* Package commit: 972b68
* Julia commit: ac5cc9
* Julia command flags: None
* Environment variables: None

## Results
Below is a table of this job's results, obtained by running the benchmarks.
The values listed in the `ID` column have the structure `[parent_group, child_group, ..., key]`, and can be used to
index into the BaseBenchmarks suite to retrieve the corresponding benchmarks.
The percentages accompanying time and memory values in the below table are noise tolerances. The "true"
time/memory value for a given benchmark is expected to fall within this percentage of the reported value.
An empty cell means that the value was zero.

| ID                                                           | time            | GC time   | memory          | allocations |
|--------------------------------------------------------------|----------------:|----------:|----------------:|------------:|
| `["Modified Multi-block (3x3)", "Gray{Float32}", "256×256"]` |  73.754 ms (5%) |  8.229 ms | 147.98 MiB (1%) |     5082158 |
| `["Modified Multi-block (3x3)", "Gray{Float32}", "512×512"]` | 306.690 ms (5%) | 37.330 ms | 592.59 MiB (1%) |    20387886 |
| `["Modified Multi-block (3x3)", "Gray{N0f8}", "256×256"]`    |  73.108 ms (5%) |  7.889 ms | 156.42 MiB (1%) |     5143662 |
| `["Modified Multi-block (3x3)", "Gray{N0f8}", "512×512"]`    | 310.055 ms (5%) | 41.396 ms | 627.47 MiB (1%) |    20641902 |
| `["Modified", "Gray{Float32}", "256×256"]`                   | 615.578 μs (5%) |           |  64.16 KiB (1%) |           3 |
| `["Modified", "Gray{Float32}", "512×512"]`                   |   2.467 ms (5%) |           | 256.16 KiB (1%) |           3 |
| `["Modified", "Gray{N0f8}", "256×256"]`                      | 823.380 μs (5%) |           |  64.16 KiB (1%) |           3 |
| `["Modified", "Gray{N0f8}", "512×512"]`                      |   3.313 ms (5%) |           | 256.16 KiB (1%) |           3 |
| `["Multi-block (1x1)", "Gray{Float32}", "256×256"]`          | 177.979 μs (5%) |           |  64.22 KiB (1%) |           5 |
| `["Multi-block (1x1)", "Gray{Float32}", "512×512"]`          | 686.318 μs (5%) |           | 256.22 KiB (1%) |           5 |
| `["Multi-block (1x1)", "Gray{N0f8}", "256×256"]`             | 227.054 μs (5%) |           |  64.22 KiB (1%) |           5 |
| `["Multi-block (1x1)", "Gray{N0f8}", "512×512"]`             | 892.521 μs (5%) |           | 256.22 KiB (1%) |           5 |
| `["Multi-block (3x3)", "Gray{Float32}", "256×256"]`          |   1.309 ms (5%) |           | 320.31 KiB (1%) |          10 |
| `["Multi-block (3x3)", "Gray{Float32}", "512×512"]`          |   5.251 ms (5%) |           |   1.25 MiB (1%) |          10 |
| `["Multi-block (3x3)", "Gray{N0f8}", "256×256"]`             |   1.310 ms (5%) |           | 320.31 KiB (1%) |          10 |
| `["Multi-block (3x3)", "Gray{N0f8}", "512×512"]`             |   5.302 ms (5%) |           |   1.25 MiB (1%) |          10 |
| `["Original (mapwindow)", "Gray{Float32}", "256×256"]`       |   1.236 ms (5%) |           |   1.34 MiB (1%) |       16340 |
| `["Original (mapwindow)", "Gray{Float32}", "512×512"]`       |   4.506 ms (5%) |           |   3.69 MiB (1%) |       32724 |
| `["Original (mapwindow)", "Gray{N0f8}", "256×256"]`          |   1.303 ms (5%) |           |   1.43 MiB (1%) |       19404 |
| `["Original (mapwindow)", "Gray{N0f8}", "512×512"]`          |   4.763 ms (5%) |           |   3.87 MiB (1%) |       38860 |
| `["Original", "Gray{Float32}", "256×256"]`                   | 174.637 μs (5%) |           |  64.16 KiB (1%) |           3 |
| `["Original", "Gray{Float32}", "512×512"]`                   | 686.605 μs (5%) |           | 256.16 KiB (1%) |           3 |
| `["Original", "Gray{N0f8}", "256×256"]`                      | 227.021 μs (5%) |           |  64.16 KiB (1%) |           3 |
| `["Original", "Gray{N0f8}", "512×512"]`                      | 892.152 μs (5%) |           | 256.16 KiB (1%) |           3 |

## Benchmark Group List
Here's a list of all the benchmark groups executed by this job:

- `["Modified Multi-block (3x3)", "Gray{Float32}"]`
- `["Modified Multi-block (3x3)", "Gray{N0f8}"]`
- `["Modified", "Gray{Float32}"]`
- `["Modified", "Gray{N0f8}"]`
- `["Multi-block (1x1)", "Gray{Float32}"]`
- `["Multi-block (1x1)", "Gray{N0f8}"]`
- `["Multi-block (3x3)", "Gray{Float32}"]`
- `["Multi-block (3x3)", "Gray{N0f8}"]`
- `["Original (mapwindow)", "Gray{Float32}"]`
- `["Original (mapwindow)", "Gray{N0f8}"]`
- `["Original", "Gray{Float32}"]`
- `["Original", "Gray{N0f8}"]`

## Julia versioninfo
```
Julia Version 1.7.1
Commit ac5cc99908 (2021-12-22 19:35 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
      Ubuntu 20.04.3 LTS
  uname: Linux 5.10.16.3-microsoft-standard-WSL2 #1 SMP Fri Apr 2 22:23:49 UTC 2021 x86_64 x86_64
  CPU: 12th Gen Intel(R) Core(TM) i9-12900K: 
                 speed         user         nice          sys         idle          irq
       #1-24  3187 MHz      54039 s        449 s      13480 s   19043491 s          0 s
       
  Memory: 31.25204849243164 GB (24743.546875 MB free)
  Uptime: 79690.41 sec
  Load Avg:  1.18  1.02  0.7
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-12.0.1 (ORCJIT, goldmont)
```
</details>

<details>
  <summary> benchmark report (master) </summary>

# Benchmark Report for *LocalBinaryPatterns*

## Job Properties
* Time of benchmark: 30 Dec 2021 - 12:0
* Package commit: 5b492e
* Julia commit: ac5cc9
* Julia command flags: None
* Environment variables: None

## Results
Below is a table of this job's results, obtained by running the benchmarks.
The values listed in the `ID` column have the structure `[parent_group, child_group, ..., key]`, and can be used to
index into the BaseBenchmarks suite to retrieve the corresponding benchmarks.
The percentages accompanying time and memory values in the below table are noise tolerances. The "true"
time/memory value for a given benchmark is expected to fall within this percentage of the reported value.
An empty cell means that the value was zero.

| ID                                                     | time            | GC time | memory          | allocations |
|--------------------------------------------------------|----------------:|--------:|----------------:|------------:|
| `["Multi-block (1x1)", "Gray{Float32}", "256×256"]`    |   1.224 ms (5%) |         | 320.25 KiB (1%) |           8 |
| `["Multi-block (1x1)", "Gray{Float32}", "512×512"]`    |   4.912 ms (5%) |         |   1.25 MiB (1%) |           8 |
| `["Multi-block (1x1)", "Gray{N0f8}", "256×256"]`       |   1.234 ms (5%) |         | 320.25 KiB (1%) |           8 |
| `["Multi-block (1x1)", "Gray{N0f8}", "512×512"]`       |   4.945 ms (5%) |         |   1.25 MiB (1%) |           8 |
| `["Multi-block (3x3)", "Gray{Float32}", "256×256"]`    |   1.265 ms (5%) |         | 320.25 KiB (1%) |           8 |
| `["Multi-block (3x3)", "Gray{Float32}", "512×512"]`    |   4.990 ms (5%) |         |   1.25 MiB (1%) |           8 |
| `["Multi-block (3x3)", "Gray{N0f8}", "256×256"]`       |   1.254 ms (5%) |         | 320.25 KiB (1%) |           8 |
| `["Multi-block (3x3)", "Gray{N0f8}", "512×512"]`       |   4.995 ms (5%) |         |   1.25 MiB (1%) |           8 |
| `["Multi-block (5x5)", "Gray{Float32}", "256×256"]`    |   1.265 ms (5%) |         | 320.25 KiB (1%) |           8 |
| `["Multi-block (5x5)", "Gray{Float32}", "512×512"]`    |   5.001 ms (5%) |         |   1.25 MiB (1%) |           8 |
| `["Multi-block (5x5)", "Gray{N0f8}", "256×256"]`       |   1.274 ms (5%) |         | 320.25 KiB (1%) |           8 |
| `["Multi-block (5x5)", "Gray{N0f8}", "512×512"]`       |   5.050 ms (5%) |         |   1.25 MiB (1%) |           8 |
| `["Original (mapwindow)", "Gray{Float32}", "256×256"]` |   1.248 ms (5%) |         |   1.34 MiB (1%) |       16340 |
| `["Original (mapwindow)", "Gray{Float32}", "512×512"]` |   4.543 ms (5%) |         |   3.69 MiB (1%) |       32724 |
| `["Original (mapwindow)", "Gray{N0f8}", "256×256"]`    |   1.314 ms (5%) |         |   1.43 MiB (1%) |       19404 |
| `["Original (mapwindow)", "Gray{N0f8}", "512×512"]`    |   4.776 ms (5%) |         |   3.87 MiB (1%) |       38860 |
| `["Original", "Gray{Float32}", "256×256"]`             | 331.874 μs (5%) |         |  64.16 KiB (1%) |           3 |
| `["Original", "Gray{Float32}", "512×512"]`             |   1.330 ms (5%) |         | 256.16 KiB (1%) |           3 |
| `["Original", "Gray{N0f8}", "256×256"]`                | 323.865 μs (5%) |         |  64.16 KiB (1%) |           3 |
| `["Original", "Gray{N0f8}", "512×512"]`                |   1.297 ms (5%) |         | 256.16 KiB (1%) |           3 |

## Benchmark Group List
Here's a list of all the benchmark groups executed by this job:

- `["Multi-block (1x1)", "Gray{Float32}"]`
- `["Multi-block (1x1)", "Gray{N0f8}"]`
- `["Multi-block (3x3)", "Gray{Float32}"]`
- `["Multi-block (3x3)", "Gray{N0f8}"]`
- `["Multi-block (5x5)", "Gray{Float32}"]`
- `["Multi-block (5x5)", "Gray{N0f8}"]`
- `["Original (mapwindow)", "Gray{Float32}"]`
- `["Original (mapwindow)", "Gray{N0f8}"]`
- `["Original", "Gray{Float32}"]`
- `["Original", "Gray{N0f8}"]`

## Julia versioninfo
```
Julia Version 1.7.1
Commit ac5cc99908 (2021-12-22 19:35 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
      Ubuntu 20.04.3 LTS
  uname: Linux 5.10.16.3-microsoft-standard-WSL2 #1 SMP Fri Apr 2 22:23:49 UTC 2021 x86_64 x86_64
  CPU: 12th Gen Intel(R) Core(TM) i9-12900K: 
                 speed         user         nice          sys         idle          irq
       #1-24  3187 MHz      22962 s        421 s       5898 s    4455205 s          0 s
       
  Memory: 31.25204849243164 GB (25637.9921875 MB free)
  Uptime: 18700.85 sec
  Load Avg:  1.07  0.82  0.42
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-12.0.1 (ORCJIT, goldmont)
```
</details>